mcp-servers/telegram-bot/requirements.txt
mcp
requests
python-dotenv==1.2.2
